backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
sqlalchemy
piper-tts
openai
python-json-logger
pytest
httpx
